# postgres_execution_package/api_spec/openapi.yaml
openapi: 3.0.3
info:
  title: منصة قطع غيار السيارات - Auto Parts Marketplace API
  version: "1.9.0"
  description: >
    مواصفة API تنفيذية، تغطي 15 عملية (Operations) عبر 14 مسارًا (Paths):
    الواجهات التمثيلية الخمس الأصلية الموثّقة بالتفصيل في التصميم التفصيلي
    (الحزمة 2)، بالإضافة إلى 5 عمليات خدمة Auth (اعتماد Contract Extension &
    Implementation Plan): تسجيل خروج، عرض/إضافة/إزالة وسائل الهوية، وعرض
    مزوّدي الهوية المفعَّلين — فتصبح عمليات Auth تحديدًا 6 عمليات شاملة
    /auth/login الأصلية. وبالإضافة أيضًا إلى 5 عمليات خدمة PCT (الكتالوج):
    اقتراح قطعة، عرضها، اعتمادها (مدير النظام فقط، REQ-PCT-002)، إضافة اسم
    مترجَم، وتسجيل رقم OEM. وبالإضافة إلى 7 عمليات خدمة VCT (كتالوج
    السيارات): اقتراح واعتماد شركة مصنّعة (مدير النظام فقط، REQ-VCT-002)،
    عرضها، اقتراح موديل تحت شركة معتمَدة (REQ-VCT-003)، إضافة جيل، إضافة
    فئة سيارة، وعرض فئة سيارة. وبالإضافة إلى 4 عمليات خدمة CMP (التوافق):
    إنشاء سجل توافق بين قطعة معتمَدة وفئة صالحة (مدير النظام فقط،
    REQ-CMP-001/002)، عرض سجل، عرض كل سجلات قطعة، وأرشفته (مدير النظام
    فقط، REQ-CMP-003). وبالإضافة أيضًا إلى خدمتَي Store وInventory: إنشاء/
    عرض متجر، تغيير حالته (مدير نظام أو مشرف، REQ-STR-004)، نقل ملكيته
    (مدير نظام حصريًا، REQ-STR-006)؛ وعرض/تحديث كمية/تسعير/إخفاء/إظهار/
    أرشفة عنصر مخزون (البائع المالك فقط، REQ-STR-019) إلى جانب /inventory-items
    الأصلية (إنشاء العنصر، مع Idempotency-Key مطلوب). جميع مسارات هذا
...
    كل عملية موسومة بامتداد x-traceability
...
  x-governance: >
...
servers:
...
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        جلسة مستخدم صادرة عن وحدة IAM بعد تسجيل دخول ناجح (REQ-IAM-003)،
        مدعومة بجدول iam.sessions (CR-013). تُسلَّم عبر Set-Cookie بخصائص
        HttpOnly (لا يقرأها JavaScript) وSecure (HTTPS فقط) وSameSite=Lax
        كحد أدنى. يُخزَّن على الخادم بصمة SHA-256 للتوكن فقط، لا التوكن
        الخام (اتساقًا مع REQ-SEC-002). غير مطلوبة لعمليات الوصول العام مثل
        تسجيل الدخول نفسه أو البحث أو عرض مزوّدي الهوية المفعَّلين.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        معرّف تتبع اختياري يرفقه المستهلك؛ إن غاب يُنشئ النظام معرّفًا جديدًا
        عند نقطة الدخول (SAD الحزمة 4، DD الحزمة 2 القسم 2.4).
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      description: نموذج الخطأ الموحّد المعتمد لجميع واجهات API (DD الحزمة 2، القسم 2.1/2.2)
      required: [correlation_id, error_code, message]
      properties:
        correlation_id:
          type: string
          format: uuid
        error_code:
          type: string
          example: "PART_NOT_APPROVED"
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
...
    InventoryItemCreateRequest:
      type: object
      required: [catalog_part_ref_id, condition_ref_id, pricing_mode, quantity]
      properties:
        catalog_part_ref_id:
          type: string
          format: uuid
        condition_ref_id:
          type: string
          format: uuid
        pricing_mode:
          type: string
          enum: [fixed_price, contact_for_price]
        price_amount:
          type: number
          format: double
        price_currency:
          type: string
          minLength: 3
          maxLength: 3
        quantity:
          type: integer
          minimum: 0
        photo_asset_refs:
          type: array
          items:
            type: string
    InventoryItemCreateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_code:
          type: string
        status:
          type: string
...
    InventoryItemFullResponse:
      type: object
      description: العرض الكامل لعنصر المخزون (يختلف عن InventoryItemCreateResponse المختصَرة عند الإنشاء)
      properties:
        id: { type: string, format: uuid }
        store_id: { type: string, format: uuid }
        catalog_part_ref_id: { type: string, format: uuid }
        condition_ref_id: { type: string, format: uuid }
        pricing_mode: { type: string, enum: [fixed_price, contact_for_price] }
        quantity: { type: integer }
        price_amount: { type: number, nullable: true }
        price_currency: { type: string, nullable: true }
        status: { type: string, enum: [active, out_of_stock, hidden, archived] }
    QuantityUpdateRequest:
      type: object
      required: [new_quantity]
      properties:
        new_quantity: { type: integer, minimum: 0 }
    PricingUpdateRequest:
      type: object
      required: [pricing_mode]
      properties:
        pricing_mode: { type: string, enum: [fixed_price, contact_for_price] }
        price_amount: { type: number, nullable: true }
        price_currency: { type: string, nullable: true }
...
paths:
...
  /api/v1/inventory/items/{itemId}:
    get:
      operationId: getInventoryItem
      summary: عرض عنصر مخزون كاملًا بمعرّفه
      x-traceability: { requirements: [REQ-STR-009], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: بيانات العنصر, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/inventory/items/{itemId}/quantity:
    patch:
      operationId: updateInventoryItemQuantity
      summary: تحديث كمية عنصر مخزون (البائع المالك فقط، REQ-STR-019)
      x-traceability: { requirements: [REQ-STR-016, REQ-STR-017, REQ-STR-019], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuantityUpdateRequest' }
      responses:
        '200': { description: تحدَّثت الكمية (وربما الحالة تلقائيًا), content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '403': { description: 'ليس البائع المالك (REQ-STR-019)', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: العنصر مؤرشَف، لا تعديل مسموح, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/inventory/items/{itemId}/pricing:
    patch:
      operationId: updateInventoryItemPricing
      summary: تحديث سياسة تسعير عنصر مخزون (البائع المالك فقط، REQ-STR-019)
      x-traceability: { requirements: [REQ-STR-012, REQ-STR-019], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PricingUpdateRequest' }
      responses:
        '200': { description: تحدَّثت السياسة, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '400': { description: سياسة تسعير غير صالحة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: ليس البائع المالك, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: العنصر مؤرشَف, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/inventory/items/{itemId}/hide:
    post:
      operationId: hideInventoryItem
      summary: إخفاء عنصر مخزون (البائع المالك فقط)
      x-traceability: { requirements: [REQ-STR-017, REQ-STR-019], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: أصبح مخفيًا, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '403': { description: ليس البائع المالك, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: العنصر مؤرشَف, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/inventory/items/{itemId}/unhide:
    post:
      operationId: unhideInventoryItem
      summary: إظهار عنصر مخزون مخفي (البائع المالك فقط)
      x-traceability: { requirements: [REQ-STR-017, REQ-STR-019], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: أصبح ظاهرًا من جديد, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '403': { description: ليس البائع المالك, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/inventory/items/{itemId}/archive:
    post:
      operationId: archiveInventoryItem
      summary: أرشفة عنصر مخزون (Soft Delete، البائع المالك فقط)
      x-traceability: { requirements: [REQ-STR-019], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: أصبح مؤرشَفًا, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryItemFullResponse' } } } }
        '403': { description: ليس البائع المالك, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: مؤرشَف بالفعل, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /api/v1/inventory-items:
    post:
      operationId: createInventoryItem
      summary: إنشاء عنصر مخزون بالإشارة لقطعة كتالوج معتمدة
      x-traceability:
        requirements: [REQ-STR-009, REQ-STR-012, REQ-STR-017, REQ-STR-020, REQ-STR-021]
        decisions: [ADR-011]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: Idempotency-Key
          in: header
          required: true
          description: >
            مطلوب (required) لهذه العملية تحديدًا (DD الحزمة 2، القسم 2.2)؛
            إعادة إرسال نفس الطلب بنفس المفتاح تُعيد النتيجة الأصلية دون
            إنشاء عنصر مخزون مكرَّر (Migration 025، sys.idempotency_keys).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemCreateRequest'
      responses:
        '201':
          description: تم إنشاء عنصر المخزون بنجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemCreateResponse'
        '400':
          description: سياسة تسعير غير صالحة، أو كمية سالبة (error_code=INVALID_PRICING/INVALID_QUANTITY)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'لا متجر نشط مملوك لهذا المستخدم (error_code=STORE_NOT_ACTIVE_OR_NOT_OWNED)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'قطعة الكتالوج غير معتمدة (error_code=PART_NOT_APPROVED)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...